Update log_pending_hook in main.c, fix log suspending in freerots task.
This patch has no effect alone. In the following commis, NRF_LOG_INFO is
used in the context of freertos task. Without this patch, log does not
work properly in freertos task (though it works in main).

I have no idea what the patch does. The code is copied from the
following discussion in Nordic community:

https://devzone.nordicsemi.com/f/nordic-q-a/91799/free-rtos-log-suspend-17-1/385406